dashboard/requirements.txt
streamlit==1.32.2
pandas==2.1.4
psycopg2-binary==2.9.9
plotly==5.18.0
sqlalchemy==2.0.25
